refactor: Remove Service interface
- 현재는 interface를 사용하지 않으므로 삭제하였습니다.
