+++ src/main/java/com/codessquad/qna/service/QuestionService.java
@@ -1,20 +1,64 @@
 package com.codessquad.qna.service;
 
+import com.codessquad.qna.domain.Answer;
 import com.codessquad.qna.domain.Question;
+import com.codessquad.qna.domain.QuestionRepository;
 import com.codessquad.qna.domain.User;
+import com.codessquad.qna.exception.ForbiddenException;
+import com.codessquad.qna.exception.NotFoundException;
+import com.codessquad.qna.exception.UnauthorizedAccessException;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
-public interface QuestionService {
-    List<Question> getQuestionList();
+@Service
+public class QuestionService {
 
-    void registerQuestion(Question question, User loginUser);
+    private final QuestionRepository questionRepository;
 
-    Question getQuestionById(Long id);
+    public QuestionService(QuestionRepository questionRepository) {
+        this.questionRepository = questionRepository;
+    }
 
-    Question getQuestionWithAuthentication(Long id, User loginUser);
+    public List<Question> getQuestionList() {
+        return questionRepository.findAllByDeleted(false);
+    }
 
-    void updateQuestion(Long id, User loginUser, Question updatedQuestion);
+    public void registerQuestion(Question question, User loginUser) {
+        question.setWriter(loginUser);
+        questionRepository.save(question);
+    }
 
-    void deleteQuestion(Long id, User loginUser);
+    public Question getQuestionById(Long id) {
+        return questionRepository.findByIdAndDeleted(id, false).orElseThrow(NotFoundException::new);
+    }
+
+    public Question getQuestionWithAuthentication(Long id, User loginUser) {
+        return questionRepository.findById(id)
+                .filter(q -> q.isWriter(loginUser))
+                .orElseThrow(() -> new UnauthorizedAccessException("다른 사람의 질문을 수정하거나 삭제할 수 없습니다."));
+    }
+
+    public void updateQuestion(Long id, User loginUser, Question updatedQuestion) {
+        Question question = getQuestionWithAuthentication(id, loginUser);
+        question.updateContents(updatedQuestion);
+        questionRepository.save(question);
+    }
+
+    @Transactional
+    public void deleteQuestion(Long id, User loginUser) {
+        Question question = getQuestionWithAuthentication(id, loginUser);
+        List<Answer> answers = question.getAnswers();
+        for (Answer answer : answers) {
+            if (!answer.matchesWriter(loginUser)) {
+                throw new ForbiddenException("다른 사람의 답변이 존재하여 질문을 삭제할 수 없습니다.");
+            }
+        }
+        for (Answer answer : answers) {
+            answer.delete();
+        }
+        question.delete();
+        questionRepository.save(question);
+    }
 }
